+++ src/main/java/org/eclipse/keyple/core/util/json/DateJsonSerializer.java
@@ -35,7 +35,8 @@ public class DateJsonSerializer implements JsonSerializer<Date> {
    * @since 2.2.0
    */
   public DateJsonSerializer() {
-    sdf = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSSXXX");
+    // We're using "Z" instead of "XXX" due to compatibility issues.
+    sdf = new SimpleDateFormat("yyyy-MM-dd'T'HH:mm:ss.SSSZ");
     sdf.setTimeZone(TimeZone.getTimeZone("CET"));
   }
 
@@ -46,6 +47,14 @@ public class DateJsonSerializer implements JsonSerializer<Date> {
    */
   @Override
   public JsonElement serialize(Date data, Type typeOfSrc, JsonSerializationContext context) {
-    return new JsonPrimitive(sdf.format(data));
+    String formatted = sdf.format(data);
+
+    // We are manually inserting a colon in the timezone offset to have a ISO 8601 format and
+    // maintain compatibility with older versions of Java and Android that do not support the "XXX"
+    // timezone marker.
+    StringBuilder sb = new StringBuilder(formatted);
+    sb.insert(formatted.length() - 2, ':');
+    formatted = sb.toString();
+    return new JsonPrimitive(formatted);
   }
 }